+++ lib/src/viewer.dart
@@ -4,22 +4,40 @@ import 'package:flutter/material.dart';
 
 import 'terms_reader.dart';
 
-typedef TermsBuilder<T> = Widget Function(
+typedef TermsBuilder = Widget? Function(
   BuildContext context,
-  T data,
+  TermsData data,
   int index,
 );
 
-typedef TermsOrderTextBuilder = String Function(
-  String style,
+typedef TermsInlineBuilder = InlineSpan? Function(
+  TermsSpan data,
+  TextStyle style,
+  int index,
+);
+
+typedef TermsOrderAlignmentBuilder = CrossAxisAlignment? Function(
+  int position,
+);
+typedef TermsOrderTextBuilder = String? Function(
+  TermsData data,
+  int index,
+);
+typedef TermsStyleBuilder = TextStyle? Function(
+  TermsData data,
+  TextStyle style,
   int index,
 );
 
 class TermsViewer extends StatelessWidget {
   final Terms data;
-  final TermsBuilder<TermsData>? titleBuilder;
-  final TermsBuilder<TermsData>? subtitleBuilder;
-  final TermsBuilder<String>? orderBuilder;
+  final TermsInlineBuilder? titleBuilder;
+  final TermsStyleBuilder? titleStyleBuilder;
+  final TermsInlineBuilder? textBuilder;
+  final TermsStyleBuilder? textStyleBuilder;
+  final TermsBuilder? orderBuilder;
+  final TermsOrderAlignmentBuilder? orderAlignmentBuilder;
+  final TermsStyleBuilder? orderStyleBuilder;
   final TermsOrderTextBuilder? orderTextBuilder;
   final double orderSpacingWidthFactor;
   final double? orderInnerSpace;
@@ -28,26 +46,37 @@ class TermsViewer extends StatelessWidget {
     super.key,
     required this.data,
     this.titleBuilder,
-    this.subtitleBuilder,
+    this.titleStyleBuilder,
+    this.textBuilder,
+    this.textStyleBuilder,
     this.orderBuilder,
+    this.orderStyleBuilder,
     this.orderTextBuilder,
     this.orderInnerSpace,
     this.orderSpacingWidthFactor = 1,
+    this.orderAlignmentBuilder,
   });
 
   @override
   Widget build(BuildContext context) {
-    int index = 0;
+    int index = -1;
     final children = data.contents;
     if (children.isEmpty) return const SizedBox();
     return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
       children: List.generate(children.length, (i) {
         final data = children[i];
         if (data.isSequenceOrder) index++;
-        return TermsBody(
+        return _Child(
           index: index,
           data: data,
+          titleBuilder: titleBuilder,
+          titleStyleBuilder: titleStyleBuilder,
+          textBuilder: textBuilder,
+          textStyleBuilder: textStyleBuilder,
+          orderAlignmentBuilder: orderAlignmentBuilder,
           orderBuilder: orderBuilder,
+          orderStyleBuilder: orderStyleBuilder,
           orderInnerSpace: orderInnerSpace,
           orderSpacingWidthFactor: orderSpacingWidthFactor,
           orderTextBuilder: orderTextBuilder,
@@ -57,19 +86,31 @@ class TermsViewer extends StatelessWidget {
   }
 }
 
-class TermsBody extends StatelessWidget {
+class _Child extends StatelessWidget {
   final int index;
   final TermsData data;
-  final TermsBuilder<String>? orderBuilder;
+
+  final TermsInlineBuilder? titleBuilder;
+  final TermsStyleBuilder? titleStyleBuilder;
+  final TermsInlineBuilder? textBuilder;
+  final TermsStyleBuilder? textStyleBuilder;
+  final TermsBuilder? orderBuilder;
+  final TermsOrderAlignmentBuilder? orderAlignmentBuilder;
+  final TermsStyleBuilder? orderStyleBuilder;
   final TermsOrderTextBuilder? orderTextBuilder;
   final double orderSpacingWidthFactor;
   final double? orderInnerSpace;
 
-  const TermsBody({
-    super.key,
-    this.index = -1,
+  const _Child({
+    required this.index,
     required this.data,
+    this.titleBuilder,
+    this.titleStyleBuilder,
+    this.textBuilder,
+    this.textStyleBuilder,
     this.orderBuilder,
+    this.orderAlignmentBuilder,
+    this.orderStyleBuilder,
     this.orderTextBuilder,
     this.orderSpacingWidthFactor = 1,
     this.orderInnerSpace,
@@ -78,12 +119,12 @@ class TermsBody extends StatelessWidget {
   String get _orderText {
     if (data.orderStyle.isEmpty) return '';
     if (!data.isSequenceOrder) return data.orderStyle;
-    return (orderTextBuilder ?? (_, __) => "($index)")(data.orderStyle, index);
+    return orderTextBuilder?.call(data, index) ?? "${index + 1}.";
   }
 
   @override
   Widget build(BuildContext context) {
-    int innerIndex = 0;
+    int innerIndex = -1;
     final position = data.position;
     final title = data.title;
     final text = data.text;
@@ -91,15 +132,24 @@ class TermsBody extends StatelessWidget {
 
     final orderText = _orderText;
 
-    const mStyle = TextStyle(fontSize: 16);
+    const style = TextStyle(fontSize: 16);
 
-    final mOrderStyle = mStyle.copyWith(
-      color: data.orderColor,
+    final mStyle = textStyleBuilder?.call(data, style, index) ?? style;
+
+    final ts = style.copyWith(
+      fontWeight: position <= 2 ? FontWeight.bold : null,
+    );
+
+    final mTitleStyle = titleStyleBuilder?.call(data, ts, index) ?? ts;
+
+    final os = mTitleStyle.copyWith(
       fontSize: data.orderFontSize,
       fontStyle: data.orderFontStyle,
       fontWeight: data.orderFontWeight,
       decoration: data.orderTextDecoration,
     );
+
+    final mOrderStyle = orderStyleBuilder?.call(data, os, index) ?? os;
 
     final mOrderSize = mOrderStyle.fontSize!;
 
@@ -110,7 +160,8 @@ class TermsBody extends StatelessWidget {
       children: [
         if (title.isNotEmpty || text.isNotEmpty) ...[
           Row(
-            crossAxisAlignment: CrossAxisAlignment.start,
+            crossAxisAlignment: orderAlignmentBuilder?.call(position) ??
+                CrossAxisAlignment.start,
             children: [
               if (position > 1)
                 SizedBox(
@@ -119,15 +170,18 @@ class TermsBody extends StatelessWidget {
                       orderSpacingWidthFactor,
                 ),
               if (orderText.isNotEmpty)
-                if (orderBuilder != null)
-                  orderBuilder!(context, orderText, index)
-                else ...[
-                  Text(
-                    orderText,
-                    style: mOrderStyle,
-                  ),
-                  SizedBox(width: mOrderSpace)
-                ],
+                Builder(
+                  builder: (context) {
+                    Widget child = Padding(
+                      padding: EdgeInsets.only(right: mOrderSpace),
+                      child: Text(orderText, style: mOrderStyle),
+                    );
+                    if (orderBuilder != null) {
+                      return orderBuilder!(context, data, index) ?? child;
+                    }
+                    return child;
+                  },
+                ),
               Expanded(
                 child: Text.rich(
                   TextSpan(
@@ -136,10 +190,27 @@ class TermsBody extends StatelessWidget {
                         TextSpan(
                           children: List.generate(title.length, (i) {
                             final e = title[i];
+                            InlineSpan? span;
+
+                            if (titleBuilder != null) {
+                              span = titleBuilder!(e, mTitleStyle, i);
+                            }
+
+                            if (span != null) return span;
+
+                            String mText = e.text;
+                            if (text.isNotEmpty && !mText.endsWith("\n")) {
+                              if (mText.endsWith(":")) {
+                                mText = "$mText ";
+                              }
+                              if (!mText.endsWith(": ")) {
+                                mText = "$mText: ";
+                              }
+                            }
+
                             return TextSpan(
-                              text: e.text,
-                              style: mStyle.copyWith(
-                                color: e.color,
+                              text: mText,
+                              style: mTitleStyle.copyWith(
                                 fontSize: e.fontSize,
                                 fontStyle: e.fontStyle,
                                 fontWeight: e.fontWeight,
@@ -152,10 +223,18 @@ class TermsBody extends StatelessWidget {
                         TextSpan(
                           children: List.generate(text.length, (i) {
                             final e = text[i];
+
+                            InlineSpan? span;
+
+                            if (textBuilder != null) {
+                              span = textBuilder!(e, mStyle, i);
+                            }
+
+                            if (span != null) return span;
+
                             return TextSpan(
                               text: e.text,
                               style: mStyle.copyWith(
-                                color: e.color,
                                 fontSize: e.fontSize,
                                 fontStyle: e.fontStyle,
                                 fontWeight: e.fontWeight,
@@ -166,7 +245,6 @@ class TermsBody extends StatelessWidget {
                         ),
                     ],
                   ),
-                  style: mStyle,
                 ),
               ),
             ],
@@ -177,13 +255,19 @@ class TermsBody extends StatelessWidget {
           ...List.generate(children.length, (i) {
             final data = children[i].copyWith(position: position + 1);
             if (data.isSequenceOrder) innerIndex++;
-            return TermsBody(
+            return _Child(
               index: innerIndex,
               data: data,
-              orderTextBuilder: orderTextBuilder,
-              orderSpacingWidthFactor: orderSpacingWidthFactor,
-              orderInnerSpace: orderInnerSpace,
+              titleBuilder: titleBuilder,
+              titleStyleBuilder: titleStyleBuilder,
+              textBuilder: textBuilder,
+              textStyleBuilder: textStyleBuilder,
               orderBuilder: orderBuilder,
+              orderAlignmentBuilder: orderAlignmentBuilder,
+              orderStyleBuilder: orderStyleBuilder,
+              orderInnerSpace: orderInnerSpace,
+              orderSpacingWidthFactor: orderSpacingWidthFactor,
+              orderTextBuilder: orderTextBuilder,
             );
           }),
       ],